Add name and location query (#12)
* Add name and location query

This add the ability to ask the name and the location settings
for the device.  Also switches to the Apache 2.0 license and
updates the README with more info.

* Moving all intents to intent_handler()
* Sticking with new-ish Configuration since it is not used by any of
  the very core functionality.  These non-critical functions are
  wrapped in try/except blocks in case some pre-0.9 code trys to use it.

* Split intents

Attempting to lump all the location queries into the same intent vocabulary
resulted in it being too broad, so phrases like "what am I?" were being
caught by the location query handler.  Split it into two different intents.