Mask vocab overlap issue
If there is overlap in the vocabulary definitions between skills, odd
issues could happen.  This has been fixed in mycroft-core, but would
still cause problems until that code gets widely deployed.  This
change masks the issue for these older systems.

Specifically, the "Location" as used in this skill is different from
the regex "Location" entity in several other skills, such as the
default skill-weather.  So "what is the forecast for seattle" was
triggering this skill.